Fixes first deploy issue with no current dir

+++ Envoy.blade.php
@@ -61,16 +61,18 @@
 
 @task('shutdown_old_docker', $on_servers)
     @for ($i=1; $i<=$num_controllers; $i++)
-        echo 'NWWS-OI Controller - Changing directory to current directory..'
-        cd /var/www/nwwsoi-controller{{ $i }}/current/
+        if [[ -L "/var/www/nwwsoi-controller{{ $i}}/current" ]]; then
+            echo 'NWWS-OI Controller - Changing directory to current directory..'
+            cd /var/www/nwwsoi-controller{{ $i }}/current/
 
-        echo 'NWWS-OI Controller - Exporting $COMPOSE_PROJECT_NAME..'
-        export COMPOSE_PROJECT_NAME=$(cat /var/www/nwwsoi-controller{{ $i }}/COMPOSE_PROJECT_NAME)
+            echo 'NWWS-OI Controller - Exporting $COMPOSE_PROJECT_NAME..'
+            export COMPOSE_PROJECT_NAME=$(cat /var/www/nwwsoi-controller{{ $i }}/COMPOSE_PROJECT_NAME)
 
-        echo 'NWWS-OI Controller - Shutting down current Docker containers..'
-        docker compose --env-file docker.env down
+            echo 'NWWS-OI Controller - Shutting down current Docker containers..'
+            docker compose --env-file docker.env down
 
-        cd /var/www/nwwsoi-controller{{ $i }}/
+            cd /var/www/nwwsoi-controller{{ $i }}/
+        fi
 
         echo 'NWWS-OI Controller - Replace current release symlink..'
         ln -nfs /var/www/nwwsoi-controller{{ $i }}/releases/{{ $releases[$i-1] }} /var/www/nwwsoi-controller{{ $i }}/current
